+++ changelog.md
@@ -25,6 +25,7 @@
 - #1459 Cart sorting: The selected items must always come before the unselected ones.
 - Checkout: After payment confirmation on the third-party site, you will be redirected directly to the completed page if successful.
 The intermediate step via the confirmation page is omitted.
+- #1465 Display the name of the applied discount on order edit page.
 
 ### Bugfixes